requirements.txt
gitpython
openai
pandoc